files/requirements.txt
configparser
paho-mqtt
json
request
